requirements.txt
fastapi
uvicorn
google-generativeai
gspread
oauth2client
langgraph
